requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
neo4j==5.12.0
tqdm==4.66.1
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.292
google-generativeai==0.1.0
scikit-learn==1.3.0
networkx==3.1
pyvis==0.3.2
matplotlib==3.7.2
seaborn==0.12.2